This is Brian Liu. Got the app up and running and changed the marker and camera to be my birthplace. Upon tapping that marker, a comment pops up indicating that this is my birth place. made another marker for current location and currently trying to figure out how to get the latitude and longitude double variables to pass in as parameters.
This is Brian Liu. Got the app up and running and changed the marker
and camera to be my birthplace. Upon tapping that marker, a comment
pops up indicating that this is my birth place. made another marker for
current location and currently trying to figure out how to get the
latitude and longitude double variables to pass in as parameters.